Fixed bug filling notes
At the time of parsing any tag type, the fields of the object are always empty. For example, when creating a new family member, its notes field is always empty (person.Note == null). But at the time of parsing the NOTE tag, the string may not be empty, but contain a note. Therefore, it is preferable to pass data from a chunk.

Added standard test CanParseIdentCONT for verification.