Added support for FAMS (FAMILY_SPOUSE) tag
Identifies the family group in which an individual appears as a partner.
The name and abbreviation of this record are misleading: the individual need not be a spouse
in that family group. The family group record is used for all relationships, not just marriages.
Do not assume that the person is a spouse in that family group.