Changed the mechanism of taking notes (NOTE, HIST tags)
It is often necessary to enter heterogeneous information under different tags in general notes to a person. For example, DefCompany (Easy Tree) has created the 'HIST' tag, which can be processed as a regular note. To support heterogeneous formats, notes are stored in the dictionary. The dictionary key is the type of note (the type of the parent tag), and the value is a list of strings.